removed Blade2 and Replaced Blade with contents  of Blade2
